double bits in collision groups

+++ libraries/entities/src/EntityItem.cpp
@@ -1914,7 +1914,7 @@ void EntityItem::setCreated(quint64 value) {
     });
 }
 
-void EntityItem::computeCollisionGroupAndFinalMask(int16_t& group, int16_t& mask) const {
+void EntityItem::computeCollisionGroupAndFinalMask(int32_t& group, int32_t& mask) const {
     if (_collisionless) {
         group = BULLET_COLLISION_GROUP_COLLISIONLESS;
         mask = 0;
@@ -1927,7 +1927,7 @@ void EntityItem::computeCollisionGroupAndFinalMask(int16_t& group, int16_t& mask
             group = BULLET_COLLISION_GROUP_STATIC;
         }
 
-        uint8_t userMask = getCollisionMask();
+        uint16_t userMask = getCollisionMask();
 
         if ((bool)(userMask & USER_COLLISION_GROUP_MY_AVATAR) !=
                 (bool)(userMask & USER_COLLISION_GROUP_OTHER_AVATAR)) {
@@ -1941,7 +1941,7 @@ void EntityItem::computeCollisionGroupAndFinalMask(int16_t& group, int16_t& mask
         if ((bool)(_flags & Simulation::SPECIAL_FLAGS_NO_BOOTSTRAPPING)) {
             userMask &= ~USER_COLLISION_GROUP_MY_AVATAR;
         }
-        mask = Physics::getDefaultCollisionMask(group) & (int16_t)(userMask);
+        mask = Physics::getDefaultCollisionMask(group) & (int32_t)(userMask);
     }
 }
 
@@ -2795,8 +2795,8 @@ bool EntityItem::getCollisionless() const {
     return result;
 }
 
-uint8_t EntityItem::getCollisionMask() const {
-    uint8_t result;
+uint16_t EntityItem::getCollisionMask() const {
+    uint16_t result;
     withReadLock([&] {
         result = _collisionMask;
     });
